Creating new Encoding class singleton to seperate encoding methods from CommonIO methods. The singleton will keep pre-loaded mappings for object sharing vs each instance having a seperate copy.  Creating Encoding-tests branch to further test internal UTF-8 interal processing.
