+++ src/common_io.hpp
@@ -7,7 +7,7 @@
 #include <vector>
 #include <map>
 #include <stdint.h>
-#include <mutex>
+
 
 // Localized Buffer.
 typedef struct localized_buffer
@@ -35,147 +35,9 @@ class CommonIO
 
 public:
 
-    explicit CommonIO()
-        : m_escape_sequence("")
-        , m_string_buffer("")
-        , m_incoming_data("")
-        , m_line_buffer("")
-        , m_column_position(0)
-        , m_is_escape_sequence(false)
-        , m_is_new_getline(true)
-        , m_is_new_leadoff(true)
-    {
-        // Arror Keys, Hardware OA are translated to [A on the fly
-        m_sequence_map.insert(std::make_pair("[A",    "up_arrow"));
-        m_sequence_map.insert(std::make_pair("[B",    "dn_arrow"));
-        m_sequence_map.insert(std::make_pair("[C",    "rt_arrow"));
-        m_sequence_map.insert(std::make_pair("[D",    "lt_arrow"));
+    explicit CommonIO();
+    ~CommonIO();
 
-        // Hardware Keys, or Numpad.
-        m_sequence_map.insert(std::make_pair("OA",    "up_arrow"));
-        m_sequence_map.insert(std::make_pair("OB",    "dn_arrow"));
-        m_sequence_map.insert(std::make_pair("OC",    "rt_arrow"));
-        m_sequence_map.insert(std::make_pair("OD",    "lt_arrow"));
-        m_sequence_map.insert(std::make_pair("OE",    "clear"));
-        m_sequence_map.insert(std::make_pair("OF",    "end"));
-        m_sequence_map.insert(std::make_pair("OH",    "home"));
-
-        // Shift Arrow Keys
-        m_sequence_map.insert(std::make_pair("[1;2A", "shift_up_arrow"));
-        m_sequence_map.insert(std::make_pair("[1;2B", "shift_dn_arrow"));
-        m_sequence_map.insert(std::make_pair("[1;2C", "shift_rt_arrow"));
-        m_sequence_map.insert(std::make_pair("[1;2D", "shift_lt_arrow"));
-
-        // Shift TAB
-        m_sequence_map.insert(std::make_pair("[Z",    "shift_tab"));
-
-        // Function Keys ANSI
-        m_sequence_map.insert(std::make_pair("[@",    "insert"));
-        m_sequence_map.insert(std::make_pair("[H",    "home"));
-        m_sequence_map.insert(std::make_pair("[K",    "end"));
-        m_sequence_map.insert(std::make_pair("[F",    "end")); // = 0F
-        m_sequence_map.insert(std::make_pair("[V",    "pg_up"));
-        m_sequence_map.insert(std::make_pair("[U",    "pg_dn"));
-        m_sequence_map.insert(std::make_pair("[OP",   "f1"));
-        m_sequence_map.insert(std::make_pair("[OQ",   "f2"));
-        m_sequence_map.insert(std::make_pair("[OR",   "f3"));
-        m_sequence_map.insert(std::make_pair("[OS",   "f4"));
-        m_sequence_map.insert(std::make_pair("[OT",   "f5"));
-        m_sequence_map.insert(std::make_pair("[[17~", "f6"));
-        m_sequence_map.insert(std::make_pair("[[18~", "f7"));
-        m_sequence_map.insert(std::make_pair("[[19~", "f8"));
-        m_sequence_map.insert(std::make_pair("[[20~", "f9"));
-        m_sequence_map.insert(std::make_pair("[[21~", "f10"));
-        m_sequence_map.insert(std::make_pair("[[23~", "f11"));
-        m_sequence_map.insert(std::make_pair("[[24~", "f12"));
-
-        // VT-100 Putty
-        m_sequence_map.insert(std::make_pair("[1~",   "home"));
-        m_sequence_map.insert(std::make_pair("[2~",   "insert"));
-        m_sequence_map.insert(std::make_pair("[3~",   "del"));
-        m_sequence_map.insert(std::make_pair("[4~",   "end"));
-        m_sequence_map.insert(std::make_pair("[5~",   "pg_up"));
-        m_sequence_map.insert(std::make_pair("[6~",   "pg_dn"));
-        m_sequence_map.insert(std::make_pair("[OU",   "f6"));
-        m_sequence_map.insert(std::make_pair("[OV",   "f7"));
-        m_sequence_map.insert(std::make_pair("[OW",   "f8"));
-        m_sequence_map.insert(std::make_pair("[OX",   "f9"));
-        m_sequence_map.insert(std::make_pair("[OY",   "f10"));
-        m_sequence_map.insert(std::make_pair("[OZ",   "f11"));
-        m_sequence_map.insert(std::make_pair("[O[",   "f12"));
-
-        // Linux Console
-        m_sequence_map.insert(std::make_pair("[[A",   "f1"));
-        m_sequence_map.insert(std::make_pair("[[B",   "f2"));
-        m_sequence_map.insert(std::make_pair("[[C",   "f3"));
-        m_sequence_map.insert(std::make_pair("[[D",   "f4"));
-        m_sequence_map.insert(std::make_pair("[[E",   "f5"));
-
-        // SCO
-        m_sequence_map.insert(std::make_pair("[L",    "insert"));
-        m_sequence_map.insert(std::make_pair("[I",    "pg_up"));
-        m_sequence_map.insert(std::make_pair("[G",    "pg_dn"));
-
-        m_sequence_map.insert(std::make_pair("[[M",   "f1"));
-        m_sequence_map.insert(std::make_pair("[[N",   "f2"));
-        m_sequence_map.insert(std::make_pair("[[O",   "f3"));
-        m_sequence_map.insert(std::make_pair("[[P",   "f4"));
-        m_sequence_map.insert(std::make_pair("[[Q",   "f5"));
-        m_sequence_map.insert(std::make_pair("[[R",   "f6"));
-        m_sequence_map.insert(std::make_pair("[[S",   "f7"));
-        m_sequence_map.insert(std::make_pair("[[T",   "f8"));
-        m_sequence_map.insert(std::make_pair("[[U",   "f9"));
-        m_sequence_map.insert(std::make_pair("[[V",   "f10"));
-        m_sequence_map.insert(std::make_pair("[[W",   "f11"));
-        m_sequence_map.insert(std::make_pair("[[X",   "f12"));
-
-        // rxvt
-        m_sequence_map.insert(std::make_pair("[7~",   "home"));
-        m_sequence_map.insert(std::make_pair("[8~",   "end"));
-
-        // Shift Arrow Keys
-        m_sequence_map.insert(std::make_pair("[a",    "shift_up_arrow"));
-        m_sequence_map.insert(std::make_pair("[b",    "shift_dn_arrow"));
-        m_sequence_map.insert(std::make_pair("[c",    "shift_rt_arrow"));
-        m_sequence_map.insert(std::make_pair("[d",    "shift_lt_arrow"));
-        m_sequence_map.insert(std::make_pair("[e",    "shift_clear"));
-
-        // Shift Function
-        m_sequence_map.insert(std::make_pair("[2$",   "insert"));
-        m_sequence_map.insert(std::make_pair("[3$",   "del"));
-        m_sequence_map.insert(std::make_pair("[5$",   "pg_up"));
-        m_sequence_map.insert(std::make_pair("[6$",   "pg_dn"));
-        m_sequence_map.insert(std::make_pair("[7$",   "home"));
-        m_sequence_map.insert(std::make_pair("[8$",   "end"));
-
-        // Ctrl
-        m_sequence_map.insert(std::make_pair("Oa",    "ctrl_up_arrow"));
-        m_sequence_map.insert(std::make_pair("Ob",    "ctrl_dn_arrow"));
-        m_sequence_map.insert(std::make_pair("Oc",    "ctrl_rt_arrow"));
-        m_sequence_map.insert(std::make_pair("Od",    "ctrl_lt_arrow"));
-        m_sequence_map.insert(std::make_pair("Oe",    "ctrl_clear"));
-
-        // Shift Function
-        m_sequence_map.insert(std::make_pair("[2^",   "ctrl_insert"));
-        m_sequence_map.insert(std::make_pair("[3^",   "ctrl_del"));
-        m_sequence_map.insert(std::make_pair("[5^",   "ctrl_pg_up"));
-        m_sequence_map.insert(std::make_pair("[6^",   "ctrl_pg_dn"));
-        m_sequence_map.insert(std::make_pair("[7^",   "ctrl_home"));
-        m_sequence_map.insert(std::make_pair("[8^",   "ctrl_end"));
-    }
-
-    ~CommonIO()
-    {
-        std::cout << "~CommonIO" << std::endl;
-        m_sequence_map.clear();
-        m_escape_sequence.erase();
-    }
-
-    static const std::string ENCODING_TEXT_UTF8;
-    static const std::string ENCODING_TEXT_CP437;
-
-    static constexpr int ENCODE_CP437 = 0;
-    static constexpr int ENCODE_UTF8  = 1;
 
     /* This function will read the OS specific functions
      * To Determine where the executable is located.
@@ -249,22 +111,6 @@ public:
      * This is for mixed ASCII And UTF-8 Strings.
      */
     bool isDigit(const std::string &str);
-
-    // Multi-Byte to WIDE (UTF-8 to UTF-16)
-    std::wstring multibyte_to_wide(const char* mbstr);
-
-    /**
-     * @brief Used for printing multibyte (Unicode Translations)
-     * @param wide_string
-     */
-    std::string wide_to_multibyte(const std::wstring &wide_string);
-
-    /**
-     * @brief Translation from CP437 to UTF-8 MultiByte Characters
-     * @param standard_string
-     */
-    std::string translateUnicode(const std::string &standard_string);
-
 
     /**
     * @brief Return the Escape Sequence
@@ -454,6 +300,5 @@ private:
     // Parameterized ESC Sequcnes Translations.
     std::map<std::string, std::string> m_sequence_map;
 
-    mutable std::mutex m;
 };
 #endif
