+++ win32/Makefile
@@ -1,8 +1,8 @@
 .PHONY: clean All
 
 All:
-	@echo "----------Building project:[ xrm-unittest - Debug ]----------"
-	@cd "C:\Users\Blue\Desktop\Oblivion2-XRM\xrm-unittest" && "$(MAKE)" -f  "xrm-unittest.mk"
+	@echo "----------Building project:[ xrm-server - Debug ]----------"
+	@"$(MAKE)" -f  "xrm-server.mk"
 clean:
-	@echo "----------Cleaning project:[ xrm-unittest - Debug ]----------"
-	@cd "C:\Users\Blue\Desktop\Oblivion2-XRM\xrm-unittest" && "$(MAKE)" -f  "xrm-unittest.mk" clean
+	@echo "----------Cleaning project:[ xrm-server - Debug ]----------"
+	@"$(MAKE)" -f  "xrm-server.mk" clean
