+++ win32/xrm-server.txt
@@ -1,2 +1,2 @@
-./Debug/up_src_data-app_oneliners_dao.cpp.o ./Debug/up_src_data-app_bbs_list_dao.cpp.o ./Debug/up_src_forms_form_system_config.cpp.o ./Debug/up_src_mods_mod_signup.cpp.o ./Debug/up_src_mods_mod_menu_editor.cpp.o ./Debug/up_src_mods_mod_user_editor.cpp.o ./Debug/up_src_mods_mod_system_config.cpp.o ./Debug/up_src_mods_mod_message_editor.cpp.o ./Debug/up_src_mods_mod_file_lister.cpp.o ./Debug/up_src_mods_mod_level_editor.cpp.o ./Debug/up_src_mods_mod_file_area_editor.cpp.o ./Debug/up_src_mods_mod_message_base_editor.cpp.o ./Debug/up_src_mods_mod_file_editor.cpp.o ./Debug/up_src_data-sys_db_startup.cpp.o ./Debug/up_src_data-sys_menu_stats_dao.cpp.o ./Debug/up_src_data-sys_access_level_dao.cpp.o ./Debug/up_src_data-sys_protocol_dao.cpp.o ./Debug/up_src_data-sys_network_dao.cpp.o ./Debug/up_src_data-sys_message_area_dao.cpp.o ./Debug/up_src_data-sys_file_area_dao.cpp.o ./Debug/up_src_communicator.cpp.o ./Debug/up_src_session_data.cpp.o ./Debug/up_src_data-sys_text_prompts_dao.cpp.o ./Debug/up_src_socket_telnet.cpp.o ./Debug/up_src_common_io.cpp.o ./Debug/up_src_encryption.cpp.o ./Debug/up_src_menu_system.cpp.o ./Debug/up_src_data-sys_menu_prompt_dao.cpp.o ./Debug/up_src_data-sys_grouping_dao.cpp.o ./Debug/up_src_session_manager.cpp.o ./Debug/up_src_mods_mod_prelogon.cpp.o ./Debug/up_src_socket_ssh.cpp.o ./Debug/up_src_mods_mod_protocol_editor.cpp.o ./Debug/up_src_ansi_processor.cpp.o ./Debug/up_src_socket_handler.cpp.o   
+./Debug/up_src_data-app_oneliners_dao.cpp.o ./Debug/up_src_data-app_bbs_list_dao.cpp.o ./Debug/up_src_forms_form_system_config.cpp.o ./Debug/up_src_mods_mod_signup.cpp.o ./Debug/up_src_mods_mod_menu_editor.cpp.o ./Debug/up_src_mods_mod_user_editor.cpp.o ./Debug/up_src_mods_mod_system_config.cpp.o ./Debug/up_src_mods_mod_message_editor.cpp.o ./Debug/up_src_mods_mod_file_lister.cpp.o ./Debug/up_src_mods_mod_level_editor.cpp.o ./Debug/up_src_mods_mod_file_area_editor.cpp.o ./Debug/up_src_mods_mod_message_base_editor.cpp.o ./Debug/up_src_mods_mod_file_editor.cpp.o ./Debug/up_src_data-sys_db_startup.cpp.o ./Debug/up_src_data-sys_menu_stats_dao.cpp.o ./Debug/up_src_data-sys_access_level_dao.cpp.o ./Debug/up_src_data-sys_protocol_dao.cpp.o ./Debug/up_src_data-sys_network_dao.cpp.o ./Debug/up_src_data-sys_message_area_dao.cpp.o ./Debug/up_src_data-sys_file_area_dao.cpp.o ./Debug/up_src_communicator.cpp.o ./Debug/up_src_session_data.cpp.o ./Debug/up_src_data-sys_text_prompts_dao.cpp.o ./Debug/up_src_socket_telnet.cpp.o ./Debug/up_src_common_io.cpp.o ./Debug/up_src_encryption.cpp.o ./Debug/up_src_menu_system.cpp.o ./Debug/up_src_data-sys_menu_prompt_dao.cpp.o ./Debug/up_src_data-sys_grouping_dao.cpp.o ./Debug/up_src_session_manager.cpp.o ./Debug/up_src_mods_mod_prelogon.cpp.o ./Debug/up_src_socket_ssh.cpp.o ./Debug/up_src_encoding.cpp.o ./Debug/up_src_mods_mod_protocol_editor.cpp.o ./Debug/up_src_ansi_processor.cpp.o ./Debug/up_src_socket_handler.cpp.o   
 ./Debug/up_src_data-sys_config_dao.cpp.o ./Debug/up_src_io_service.cpp.o ./Debug/up_src_access_condition.cpp.o ./Debug/up_src_telnet_decoder.cpp.o ./Debug/up_src_sdl2_net_SDLnet.cpp.o ./Debug/up_src_sdl2_net_SDLnetselect.cpp.o ./Debug/up_src_process_win.cpp.o ./Debug/up_src_state_manager.cpp.o ./Debug/up_src_data-sys_menu_dao.cpp.o ./Debug/up_src_mods_mod_form_manager.cpp.o ./Debug/up_src_main.cpp.o ./Debug/up_src_menu_base.cpp.o ./Debug/up_src_sdl2_net_SDLnetTCP.cpp.o ./Debug/up_src_sdl2_net_SDLnetUDP.cpp.o ./Debug/up_src_form_manager.cpp.o ./Debug/up_src_data-sys_security_dao.cpp.o ./Debug/up_src_session_io.cpp.o ./Debug/up_src_data-sys_users_dao.cpp.o ./Debug/up_src_mods_mod_logon.cpp.o ./Debug/up_src_data-sys_session_stats_dao.cpp.o ./Debug/up_src_data-sys_conference_dao.cpp.o ./Debug/up_src_mods_mod_message_reader.cpp.o
