+++ xrm-unittest/xrm-unittest.txt
@@ -1,1 +1,1 @@
-Debug/main.cpp.o Debug/common_io_ut.cpp.o Debug/session_io_ut.cpp.o Debug/ansi_processor_ut.cpp.o Debug/src_ansi_processor.cpp.o Debug/users_dao_ut.cpp.o Debug/src_session_manager.cpp.o Debug/src_common_io.cpp.o Debug/src_session_io.cpp.o Debug/access_condition_ut.cpp.o Debug/src_access_condition.cpp.o Debug/form_system_config_ut.cpp.o Debug/src_form_manager.cpp.o Debug/conference_dao_it.cpp.o Debug/file_area_dao_it.cpp.o Debug/message_area_dao_it.cpp.o Debug/oneliners_dao_it.cpp.o Debug/src_io_service.cpp.o Debug/src_communicator.cpp.o Debug/src_socket_handler.cpp.o Debug/src_socket_ssh.cpp.o Debug/src_socket_telnet.cpp.o Debug/data-sys_config_dao.cpp.o Debug/data-sys_conference_dao.cpp.o Debug/data-sys_menu_dao.cpp.o Debug/data-sys_file_area_dao.cpp.o Debug/data-sys_message_area_dao.cpp.o Debug/data-sys_access_level_dao.cpp.o Debug/data-sys_protocol_dao.cpp.o Debug/data-sys_security_dao.cpp.o Debug/data-sys_session_stats_dao.cpp.o Debug/data-sys_users_dao.cpp.o Debug/data-sys_text_prompts_dao.cpp.o Debug/forms_form_system_config.cpp.o Debug/data-app_oneliners_dao.cpp.o Debug/sdl2_net_SDLnet.cpp.o Debug/sdl2_net_SDLnetselect.cpp.o Debug/sdl2_net_SDLnetTCP.cpp.o Debug/sdl2_net_SDLnetUDP.cpp.o
+Debug/up_src_sdl2_net_SDLnet.cpp.o Debug/access_condition_ut.cpp.o Debug/up_src_access_condition.cpp.o Debug/users_dao_ut.cpp.o Debug/up_src_ansi_processor.cpp.o Debug/form_system_config_ut.cpp.o Debug/up_src_communicator.cpp.o Debug/up_src_data-sys_security_dao.cpp.o Debug/up_src_form_manager.cpp.o Debug/common_io_ut.cpp.o Debug/up_src_data-sys_message_area_dao.cpp.o Debug/up_src_socket_telnet.cpp.o Debug/session_io_ut.cpp.o Debug/up_src_data-sys_text_prompts_dao.cpp.o Debug/up_src_socket_handler.cpp.o Debug/main.cpp.o Debug/oneliners_dao_it.cpp.o Debug/conference_dao_it.cpp.o Debug/up_src_socket_ssh.cpp.o Debug/up_src_common_io.cpp.o Debug/up_src_data-sys_db_startup.cpp.o Debug/up_src_session_io.cpp.o Debug/up_src_io_service.cpp.o Debug/up_src_data-sys_access_level_dao.cpp.o Debug/file_area_dao_it.cpp.o Debug/up_src_data-sys_config_dao.cpp.o Debug/up_src_data-sys_conference_dao.cpp.o Debug/up_src_data-sys_menu_dao.cpp.o Debug/up_src_data-sys_file_area_dao.cpp.o Debug/message_area_dao_it.cpp.o Debug/ansi_processor_ut.cpp.o Debug/up_src_data-sys_users_dao.cpp.o Debug/up_src_data-sys_protocol_dao.cpp.o Debug/up_src_data-sys_session_stats_dao.cpp.o Debug/up_src_forms_form_system_config.cpp.o Debug/up_src_data-app_oneliners_dao.cpp.o Debug/up_src_session_manager.cpp.o Debug/up_src_sdl2_net_SDLnetTCP.cpp.o Debug/up_src_sdl2_net_SDLnetUDP.cpp.o Debug/up_src_sdl2_net_SDLnetselect.cpp.o
